requirements.txt
astroid==2.3.3
boto3==1.12.31
botocore==1.15.31
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
future==0.18.2
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
prettytable==0.7.2
pylint==2.4.4
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
twitter==1.18.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
